feat: compile csv file on server
