+++ class-components/src/components/flyout-component/flyout-component.tsx
@@ -4,23 +4,7 @@ import type { RootState } from '../../redux/store';
 import { unselectAll } from '../../redux/slices/selected-pokemon-slice';
 import { useRef } from 'react';
 import { useTranslations } from 'next-intl';
-
-type CsvRow = {
-  id: number;
-  description: string;
-  url: string;
-};
-
-function convertToCSV(items: CsvRow[]): string {
-  if (items.length === 0) return '';
-
-  const headers = ['id', 'description', 'url'];
-  const rows = items.map((item) =>
-    [item.id, item.description.replaceAll(',', ''), item.url].join(',')
-  );
-
-  return [headers.join(','), ...rows].join('\n');
-}
+import { compileCsvAction } from '../../app/actions/exportCsv';
 
 export default function FlyOutComponent() {
   const downloadLinkRef = useRef<HTMLAnchorElement | null>(null);
@@ -30,17 +14,18 @@ export default function FlyOutComponent() {
   const handleClick = () => {
     dispatch(unselectAll());
   };
-  const handleDownload = () => {
+  const handleDownload = async () => {
     if (!csv || csv.length === 0) return;
 
-    const csvContent = convertToCSV(csv);
-    const blob = new Blob([csvContent], { type: 'text/csv;charset=utf-8;' });
+    const csvContent = await compileCsvAction(csv);
+    const blob = new Blob([csvContent.csv], {
+      type: 'text/csv;charset=utf-8;',
+    });
     const url = URL.createObjectURL(blob);
-    const fileName = `${csv.length}_items.csv`;
 
     if (downloadLinkRef.current) {
       downloadLinkRef.current.href = url;
-      downloadLinkRef.current.download = fileName;
+      downloadLinkRef.current.download = csvContent.filename;
       downloadLinkRef.current.click();
     }
   };
